airflow/requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-amazon
minio
psycopg2-binary
boto3